requirements.txt
basicsr
diffusers
fairscale==0.4.4
fonts
font-roboto
gfpgan
gradio==3.4.1
invisible-watermark
numpy
omegaconf
piexif
Pillow
pytorch_lightning
realesrgan
scikit-image>=0.19
timm==0.4.12
transformers==4.19.2
torch
einops
jsonmerge
clean-fid
resize-right
torchdiffeq
kornia
lark
flask
flask-socketio
